address comments; split persona invocation into one class per file
