+++ src/domain/community/virtual-contributor/dto/prompt-graph-definition/prompt.graph.definition.dto.ts
