+++ src/domain/community/virtual-contributor/transformers/prompt.graph.transformer.ts
@@ -1,4 +1,4 @@
-import { PromptGraphDefinition } from '../dto/prompt-graph-definition/prompt.graph.defintion.dto';
+import { PromptGraphDefinition } from '../dto/prompt-graph-definition/prompt.graph.definition.dto';
 
 export const PromptGraphTransformer = {
   to: (value?: PromptGraphDefinition) => value,
