+++ src/domain/community/virtual-contributor/virtual.contributor.interface.ts
@@ -12,7 +12,7 @@ import { VirtualContributorInteractionMode } from '@common/enums/virtual.contrib
 import { UUID } from '@domain/common/scalars';
 import { VirtualContributorBodyOfKnowledgeType } from '@common/enums/virtual.contributor.body.of.knowledge.type';
 import { ISpace } from '@domain/space/space/space.interface';
-import { PromptGraphDefinition } from './dto/prompt-graph-definition/prompt.graph.defintion.dto';
+import { PromptGraphDefinition } from './dto/prompt-graph-definition/prompt.graph.definition.dto';
 import { IAiPersona } from '@services/ai-server/ai-persona';
 
 @ObjectType('VirtualContributor', {
