+++ src/services/ai-server/ai-persona/ai.persona.module.ts
@@ -7,7 +7,7 @@ import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/a
 import { AiPersonaEngineAdapterModule } from '../ai-persona-engine-adapter/ai.persona.engine.adapter.module';
 import { AiPersonaAuthorizationService } from './ai.persona.service.authorization';
 import { AiPersonaResolverMutations } from './ai.persona.resolver.mutations';
-import { AiPersonaResolverFields } from './ai.persona.service.resolver.fields';
+import { AiPersonaResolverFields } from './ai.persona.resolver.fields';
 import { AiPersonaExternalConfigResolverFields } from './ai.persona.external.config.resolver.fields';
 
 @Module({
