+++ src/services/ai-server/ai-persona/dto/ai.persona.invocation/invocation.operation.dto.ts
@@ -1,0 +1,11 @@
+import { registerEnumType } from '@nestjs/graphql';
+
+export enum InvocationOperation {
+  QUERY = 'query',
+  INGEST = 'ingest',
+}
+
+registerEnumType(InvocationOperation, {
+  name: 'InvocationOperation',
+  description: 'Available operations for the engine to execute.',
+});
