+++ src/services/ai-server/ai-persona/dto/index.ts
@@ -1,6 +1,10 @@
 export * from './ai.persona.dto.create';
 export * from './ai.persona.dto.update';
 export * from './ai.persona.dto.delete';
-export * from './ai.persona.invocation.dto.input';
+export * from './ai.persona.invocation/ai.persona.invocation.dto.input';
+export * from './ai.persona.invocation/result.handler.dto';
+export * from './ai.persona.invocation/room.details.dto';
+export * from './ai.persona.invocation/invocation.operation.dto';
+export * from './ai.persona.invocation/invocation.result.action.dto';
 export * from './interaction.message';
 export * from './external.config';
