+++ src/services/ai-server/prompt-graph/dto/prompt.graph.data.struct.dto.ts
@@ -1,8 +1,9 @@
-import { ObjectType, Field } from '@nestjs/graphql';
+import { ObjectType, Field, InputType } from '@nestjs/graphql';
 import { PromptGraphDataPoint } from './prompt.graph.data.point.dto';
 
+@InputType('PromptGraphDataStructInput')
 @ObjectType()
-export class PromptGrapDataStruct {
+export class PromptGraphDataStruct {
   @Field({ nullable: true })
   title?: string;
