+++ src/services/ai-server/prompt-graph/dto/prompt.graph.dto.ts
@@ -1,9 +1,10 @@
-import { ObjectType, Field } from '@nestjs/graphql';
+import { ObjectType, Field, InputType } from '@nestjs/graphql';
 
 import { PromptGraphNode } from './prompt.graph.node.dto';
 import { PromptGraphEdge } from './prompt.graph.edge.dto';
-import { PromptGrapDataStruct } from './prompt.grap.data.struct.dto';
+import { PromptGraphDataStruct } from './prompt.graph.data.struct.dto';
 
+@InputType('PromptGraphInput')
 @ObjectType()
 export class PromptGraph {
   @Field(() => [PromptGraphNode], { nullable: true })
@@ -18,6 +19,6 @@ export class PromptGraph {
   @Field({ nullable: true })
   end?: string;
 
-  @Field(() => PromptGrapDataStruct, { nullable: true })
-  state?: PromptGrapDataStruct;
+  @Field(() => PromptGraphDataStruct, { nullable: true })
+  state?: PromptGraphDataStruct;
 }
